pyproject.toml
[project]
name = "twilioopenai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "audioop-lts>=0.2.2",
    "fastapi>=0.117.1",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
    "twilio>=9.8.1",
    "uvicorn[standard]>=0.36.0",
    "websockets>=15.0.1",
    "ruff>=0.7.0",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
